First correct detections
Should go back to the `make_edges` method and check the "transitivity"
rule. Right now, I insert edge (j,i) if either j is "close" (in time and
distance) to i or if it is close in distance to i and close in time to
some k such that (k,i) is an edge.